+++ packages/components/src/components/modal/modal-styles.ts
@@ -1,12 +1,14 @@
 import type { JssStyle } from 'jss';
-import type { GetJssStyleFunction } from '../../utils';
+import type { GetJssStyleFunction, Theme } from '../../utils';
 import {
   buildResponsiveStyles,
   getCss,
   isHighContrastMode,
+  isThemeDark,
   mergeDeep,
   parseJSON,
   scrollShadowColor,
+  scrollShadowColorDark,
 } from '../../utils';
 import type { Breakpoint } from '@porsche-design-system/utilities-v2';
 import {
@@ -26,12 +28,12 @@ import {
   getThemedColors,
   hostHiddenStyles,
   hoverMediaQuery,
+  prefersColorSchemeDarkMediaQuery,
   pxToRemWithUnit,
 } from '../../styles';
 import { MODAL_Z_INDEX } from '../../constants';
 
 const mediaQueryXl = getMediaQueryMin('xl');
-const { backgroundColor, primaryColor: lightThemePrimaryColor } = getThemedColors('light');
 const { primaryColor: darkThemePrimaryColor, contrastHighColor: darkThemeContrastHighColor } = getThemedColors('dark');
 
 const transitionTimingFunction = 'cubic-bezier(.16,1,.3,1)';
@@ -95,8 +97,11 @@ export const getComponentCss = (
   isFullscreen: BreakpointCustomizable<boolean>,
   hasDismissButton: boolean,
   hasHeader: boolean,
-  hasFooter: boolean
+  hasFooter: boolean,
+  theme: Theme
 ): string => {
+  const { primaryColor, backgroundColor } = getThemedColors(theme);
+  const { primaryColor: primaryColorDark, backgroundColor: backgroundColorDark } = getThemedColors('dark');
   const isFullscreenForXlAndXxl = isFullscreenForXl(isFullscreen);
   const duration = isOpen ? '.6s' : '.2s';
   const contentPadding = '32px';
@@ -122,7 +127,7 @@ export const getComponentCss = (
               }),
           ...colorSchemeStyles,
           ...hostHiddenStyles,
-          ...getFrostedGlassBackgroundJssStyles(isOpen, duration),
+          ...getFrostedGlassBackgroundJssStyles(isOpen, duration, theme),
         }),
         overflowY: 'auto', // overrideable
       },
@@ -144,11 +149,15 @@ export const getComponentCss = (
       h2: {
         ...headingLargeStyle,
         margin: 0,
-        color: lightThemePrimaryColor,
+        color: primaryColor,
+        ...prefersColorSchemeDarkMediaQuery(theme, {
+          color: primaryColorDark,
+        }),
       },
     },
     root: mergeDeep(
       {
+        color: primaryColor, // enables color inheritance for slotted content
         position: 'relative',
         boxSizing: 'border-box',
         transform: isOpen ? 'scale3d(1,1,1)' : 'scale3d(.9,.9,1)',
@@ -166,8 +175,11 @@ export const getComponentCss = (
           pointerEvents: 'none', // fix text selection in focus state
           ...buildResponsiveStyles(isFullscreen, (fullscreenValue: boolean) => ({
             borderRadius: fullscreenValue ? 0 : '12px',
-            borderColor: fullscreenValue ? lightThemePrimaryColor : darkThemePrimaryColor,
+            borderColor: fullscreenValue ? primaryColor : darkThemePrimaryColor,
             ...getInsetJssStyle(fullscreenValue ? 0 : -4),
+            ...prefersColorSchemeDarkMediaQuery(theme, {
+              borderColor: darkThemePrimaryColor,
+            }),
           })),
         },
         '&:not(:focus-visible)::before': {
@@ -176,6 +188,10 @@ export const getComponentCss = (
         [mediaQueryXl]: {
           margin: isFullscreenForXlAndXxl ? 0 : `min(192px, 10vh) ${gridExtendedOffsetBase}`,
         },
+        ...prefersColorSchemeDarkMediaQuery(theme, {
+          color: primaryColorDark,
+          background: backgroundColorDark,
+        }),
       },
       buildResponsiveStyles(isFullscreen, getFullscreenJssStyles) as any // potentially needs to be merged with mediaQueryXl
     ),
@@ -199,10 +215,16 @@ export const getComponentCss = (
         bottom: 0,
         borderBottomLeftRadius: borderRadiusMedium,
         borderBottomRightRadius: borderRadiusMedium,
+        ...prefersColorSchemeDarkMediaQuery(theme, {
+          background: backgroundColorDark,
+        }),
       },
       [footerShadowClass]: {
-        boxShadow: `${scrollShadowColor} 0 -5px 10px`,
+        boxShadow: `${isThemeDark(theme) ? scrollShadowColorDark : scrollShadowColor} 0 -5px 10px`,
         clipPath: 'inset(-20px 0 0 0)', // crop leaking box-shadow on left and right side
+        ...prefersColorSchemeDarkMediaQuery(theme, {
+          boxShadow: `${scrollShadowColorDark} 0 -5px 10px`,
+        }),
       },
     }),
     ...(hasDismissButton && {
@@ -218,6 +240,10 @@ export const getComponentCss = (
             background: darkThemeContrastHighColor,
             borderColor: darkThemeContrastHighColor,
           },
+        }),
+        ...prefersColorSchemeDarkMediaQuery(theme, {
+          background: backgroundColorDark,
+          borderColor: backgroundColorDark,
         }),
       },
     }),