+++ packages/components/CHANGELOG.md
@@ -16,8 +16,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Added
 
+- `Popover` and `Modal` support theme dark and auto
+  ([#2789](https://github.com/porsche-design-system/porsche-design-system/pull/2789))
 - Styles: `getSkeletonStyle()` and `pds-skeleton`
   ([#2796](https://github.com/porsche-design-system/porsche-design-system/pull/2796))
+
+#### Changed
+
+- Styles: `themeDarkBackgroundShading` and `pds-theme-dark-background-shading` color
+  ([#2789](https://github.com/porsche-design-system/porsche-design-system/pull/2789))
 
 ### [3.7.0] - 2023-10-04